feat/ref: 유효성 검사 추가, 비밀번호 검증 수정
- 비밀번호를 검증하는 로직을 repository -> service로 이동
- 유효성 검사 @Valid 추가, 요구사항 lv6 수행

+++ src/main/java/com/example/schedule/repository/UserRepository.java
@@ -3,6 +3,8 @@ package com.example.schedule.repository;
 import com.example.schedule.entity.UserInfo;
 
 public interface UserRepository {
+
     int createUserInfo(UserInfo userinfo);
-    int updateUserName(int id, String updateName);
+    int updateUserName(Long id, String updateName);
+
 }
